LayoutTests/platform/ios-simulator/fast/events/touch/gesture/long-press-on-draggable-element-in-iframe-triggers-drag-expected.txt

CONSOLE MESSAGE: line 24: TypeError: undefined is not a function (evaluating 'internals.settings.setTouchDragDropEnabled(true)')
FAIL: Timed out waiting for notifyDone to be called

This test checks that a long press gesture on a draggable element in an iframe starts a drag/drop session.

Testing
FAIL
